pyproject.toml
[tool.poetry]
name = "customermanagement"
version = "0.1.0"
description = ""
authors = ["Vinodrd <vinod007rs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.1.0"
pyjwt = "^2.7.0"
pylint = "^2.17.4"
django-environ = "^0.10.0"
gunicorn = "^20.1.0"
environ = "^1.0"
django-phonenumber-field = "^7.1.0"
python-dotenv = "^1.0.0"
phonenumbers = "^8.13.15"
drf-yasg = "^1.21.6"
google-cloud-secret-manager = "^2.16.1"
django-storages = "^1.13.2"
google = "^3.0.0"
google-auth = "^2.21.0"
google-cloud = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
